+++ CHANGELOG.md
@@ -1,14 +1,32 @@
-# Changelog
+# CHANGELOG
 
 All notable changes to `shergela/validation-rule` will be documented in this file.
 
-## 1.0.1 - Friday 08, August - 2024
+## 1.0.2 | 19 August - 2024
 
+## These methods work with timezones
+### You need to provide city names in the methods.
+
+- add new `timezones()` method
+- add new `timezoneAfrica()` method
+- add new `timezoneAsia()` method
+- add new `timezoneEurope()` method
+- add new `timezoneAmerica()` method
+- add new `timezoneAntarctica()` method
+- add new `timezoneArctic()` method
+- add new `timezoneAtlantic()` method
+- add new `timezoneAustralia()` method
+- add new `timezoneIndian()` method
+- add new `timezonePacific()` method
+
+## 1.0.1 | 16 August - 2024
+
+- add new `lowercaseFirstLetter()` method
 - add new `separateIntegersByComma()` method
 - add new `separateStringsByComma()` method
 - add new `separateStringsByUnderscore()` method
 
-## 1.0.0 - Tuesday 23, July - 2024
+## 1.0.0 | 23 July - 2024
 
 - add new `size()` method
 - add new `endsWith()` method